salad-svc/service/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
python-multipart
pillow
numpy
pandas==2.1.3
prettytable==3.9.0
pytorch-lightning==2.1.2
pytorch-metric-learning==2.3.0
torchmetrics==1.2.0
xformers==0.0.23.post1
faiss-gpu==1.7.2
